HU1 Autenticacion - modificaciones necesarias para HU2 solicitudes.

+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/bancolombia/api/dto/response/UserRecordResponse.java
@@ -12,7 +12,7 @@ public record UserRecordResponse(
     String address,
     String email,
     String phone,
-    BigInteger baseSalary
+    Double baseSalary
 ) {
-    
+
 }
